requirements.txt
pytest==6.1.1
Appium-Python-Client==1.0.2
urllib3==1.26.5
